perf: sort libphonenumber-js du bundle client de la home (-50 ko gzip)
La chaîne SearchBar -> base-adresse -> barrel shared -> address.model ->
zod-primitives embarquait libphonenumber-js (39 ko gzip) et les validateurs
siret/phone dans le first-load de la home, pour un simple buildEnum et deux
bornes latitude/longitude.

- extrait buildEnum/latitude/longitude dans zod-primitives-light.ts (sans
  dépendance lourde) ; zod-primitives les ré-expose dans extensions pour ses
  ~40 importeurs existants
- routes/params.ts et models/address.model.ts basculent sur le module léger
- base-adresse.ts et search.params.utils.ts passent en imports profonds au
  lieu du barrel shared
- corrige le chemin sideEffects de shared/package.json qui pointait vers un
  fichier inexistant (zodWithOpenApi.js vs zod-with-open-api.js)

First-load JS de la home : 576 -> 526 ko gzip. Typecheck des 3 workspaces OK,
122 tests shared OK, autocomplete de lieu vérifiée en build prod local.
